requirements.txt
fastapi
uvicorn
requests
pymongo
openai
python-dotenv
geocoder 
pytz
transformers
ctransformers
accelerate
spacy
